Change file strucure
