+++ library/linalg_lib.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "matrix/matrix.h"
+#include "matrix/matrix_view.h"
 #include "matrix/sparse_linear_transformation.h"
 #include "algo/givens_qr.h"
 #include "algo/householder_qr.h"
